Put common possible common methods into parent game class
