+++ cpp/constants.h
@@ -1,0 +1,75 @@
+#ifndef CONSTANTS_HPP
+#define CONSTANTS_HPP
+
+constexpr int boringColors[] = {75, 100, 150, 200, 225, 250};
+
+const bool numberPatterns[10][5][3] = {{
+  {1, 1, 1},
+  {1, 0, 1},
+  {1, 0, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+}, {
+  {0, 1, 0},
+  {1, 1, 0},
+  {0, 1, 0},
+  {0, 1, 0},
+  {0, 1, 0},
+}, {
+  {1, 1, 1},
+  {0, 0, 1},
+  {1, 1, 1},
+  {1, 0, 0},
+  {1, 1, 1},
+}, {
+  {1, 1, 1},
+  {0, 0, 1},
+  {1, 1, 1},
+  {0, 0, 1},
+  {1, 1, 1},
+}, {
+  {1, 0, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+  {0, 0, 1},
+  {0, 0, 1},
+}, {
+  {1, 1, 1},
+  {1, 0, 0},
+  {1, 1, 1},
+  {0, 0, 1},
+  {1, 1, 1},
+}, {
+  {1, 1, 1},
+  {1, 0, 0},
+  {1, 1, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+}, {
+  {1, 1, 1},
+  {0, 0, 1},
+  {0, 0, 1},
+  {0, 1, 0},
+  {0, 1, 0},
+}, {
+  {1, 1, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+}, {
+  {1, 1, 1},
+  {1, 0, 1},
+  {1, 1, 1},
+  {0, 0, 1},
+  {1, 1, 1},
+}};
+
+struct Rect {
+  float x;
+  float y;
+  float width;
+  float height;
+};
+
+#endif
